refactor: use hooks and address pr comments
